* main:
  Vhost ratio for XGBoost provider (#994)
  Codespaces merge into main (#993)
  Container energy idle (#992)
  Monotonic and low value check (#973)
  Phase Stats are now also saved if runs fails (#972)
  RAPL energy filtering moved to provider (#979)
  fixes the GCC warnings (#985)
  Bump psycopg-pool from 3.2.3 to 3.2.4 (#982)
  Bump aiohttp from 3.11.2 to 3.11.4 (#984)
  Fixed SCI link
  Hinting UTC
  Removed duplicate tests and fixed tests
  Bump aiohttp from 3.11.0 to 3.11.2 (#981)
  Added unit test for rescale energy value
  Bump aiohttp from 3.10.10 to 3.11.0 (#980)
  (fix): Repo submission to software/add was failing on non-main branch

+++ lib/install_shared.sh
@@ -236,6 +236,9 @@ function build_binaries() {
             if [[ $(uname) == "Linux" ]] && [[ "$make_path" == *"/mach/"* ]]; then
                 continue
             fi
+            if [[ "$make_path" == *"/lmsensors/"* ]] && [[ "${install_sensors}" == false ]]; then
+                continue
+            fi
             echo "Installing $subdir/metric-provider-binary ..."
             rm -f $subdir/metric-provider-binary 2> /dev/null
             make -C $subdir
@@ -304,7 +307,6 @@ while getopts "p:a:m:nhtbisyrlc:k:e:" o; do
             ;;
         s)
             install_sensors=false
-            # currently unused
             ;;
         r)
             install_msr_tools=false
